Remove event_loop and loop fixtures
 - now requires pytestio-asyncio >= 0.25
 - use asyncio.get_running_loop() in order to access event loop
